+++ packages/helpers/CHANGELOG.md
@@ -1,7 +1,12 @@
 
+2023-03-07 (v2.1.4)
+-------------------
+
 - Support `wrapUri()` method
 - Fix: ExplodeVeridaUri regex breaking when there are no query params
 - Add explode DID helper method
+- Support deep attributes in Verida URIs
+- Add unit tests
 
 2023-02-16 (v2.1.1)
 -------------------